Add files via upload
App.py: It is a flask app python file.

Model.py: This contains code for our Machine Learning model to predict our sudoku data and testing on it.

Request.py: This uses requests module to call APIs already defined in app.py and displays the returned value.

Sudoku.html: It is html file representation of our Model.py

Sudoku_design_frame.html: This file should be present inside templates folder in our directory. It consists of an interface page build on HTML/CSS to get user inputs and show our results.

Readme.txt: It consists of 2 links. One link is for the Kaggle dataset linked to (1 Million dataset) and the other is having saved model CNN with weights.